requirements.txt
crewai 
langchain_openai 
crewai[tools]
gradio
fpdf
langchain_community
python-dotenv 